tests: Streamline testing for LRO
- Disable testing of snippets rather than having all the files with "test disabled"
- Remove the build test (which would require us to have snippet csproj files)
- Move the fakes out of the folder where we expect generated code to live
- Test against the full generated file instead of just small parts of it
