+++ Google.Api.Generator.Tests/ProtoTests/Lro/Lro.proto
@@ -6,6 +6,7 @@ import "google/api/resource.proto";
 
 package testing.lro;
 
+// LRO service to test.
 service Lro {
   option (google.api.default_host) = "lro.example.com";
