--- Google.Api.Generator.Tests/ProtoTests/Lro/Testing.Lro.GeneratedSnippets/snippet_metadata_testing.lro.json
@@ -1,1 +1,0 @@
-﻿// TEST_DISABLE
